feat: add script to cleanup stale test data
